+++ examples/WiFi/sender/sender.ino
@@ -13,7 +13,7 @@ const IPAddress subnet(255, 255, 255, 0);
 
 ArtnetWiFiSender artnet;
 const String target_ip = "192.168.1.200";
-uint32_t universe = 1;
+uint16_t universe = 1;
 
 const uint16_t size = 512;
 uint8_t data[size];
